drivers: sec_ts: Expose interfaces for lowpowermode state change
Reference the sysfs interface for setting lowpower mode.

The touchscreen should enter lowpower mode in some cases, for example,
when entering AOD mode. However, the AOD mode handling is at the
msm display module in techpack/display. So expose functions to
enable/disable lowpower mode such that other modules are able to
make use of it.

Signed-off-by: Luofan Chen <clfbbn@gmail.com>
Change-Id: I12e035774218a6d442be0afa1f8830afd207dfd9
Signed-off-by: BladeRunner-A2C <john.smith@unused.email>